完成攉龙设置界面的UI设计,完成逻辑框架,填充了一部分操作和数据修改逻辑, 接下来完成剩余的逻辑并对逻辑自洽性进行充分测试. 整理优化了界面显示逻辑, 更改了游戏初始设定为多项,以便存储几个常用设置, 比如攉龙常用4人3副和4人4副规则

+++ Assets/GameBase/Script/Common/Core/SystemSettings.cs
@@ -8,6 +8,6 @@ namespace GameBase.Common.Core
         public bool musicMute;
         public float soundVolume;
         public bool soundMute;
-        public float aiDelay;
+        public int aiDelay;
     }
 }
